backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator==2.1.0
httpx==0.27.2

# Mensajería anónima
redis==5.0.1
python-socketio==5.10.0
cryptography==41.0.7
PyJWT==2.8.0

# Tareas programadas
apscheduler==3.10.4